Overall refactoring and cleanup
