+++ Sources/StreamChatSwiftUI/ChatComposer/VoiceRecording/LockView.swift
